discord-clan-bot/requirements.txt
discord.py==2.3.2
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1
schedule==1.2.0
pytz==2023.3